requirements.txt
streamlit
PyPDF2
langchain
google-generativeai
faiss-cpu
langchain-google-genai
langchain_google_genai
python-dotenv